Device priority and buzz-removal working!
Buzzing noise that happened sometimes after connecting a new device
(e.g. HDMI) is now fixed by re-setting the thru-engine's buffer size to
its current value. Also, a sort of "device priority" is implemented so
that every time a new device is connected, the one with the highest
priority is chosen.